feat: rework error page location
