+++ frontend/src/common/pages/IssuanceErrorPage.scss
@@ -1,0 +1,11 @@
+@import "../../themes/orbcStyles";
+
+.issuance-error-page {
+    &__msg {
+        &--top {
+            margin-top: 0;
+            margin-bottom: 1.5rem;
+            display: block;
+        }
+    }
+}
