--- frontend/src/features/permits/pages/IssuanceErrorPage/IssuanceErrorPage.scss
@@ -1,8 +1,0 @@
-@use "../../../../themes/orbcStyles";
-
-.issuance-error-container {
-    .block-1 {
-        margin-top: 5px;
-        margin-bottom: 20px;
-    }
-}
